+++ backend/Models/TransactionModel.cs
@@ -28,7 +28,7 @@ namespace backend.Models
 
 		[BsonElement("amount")]
 		[JsonProperty("amount")]
-		public Decimal Amount { get; set; }
+		public int Amount { get; set; }
 
 		[BsonElement("currency")]
 		[JsonProperty("currency")]
